+++ CHANGES.md
@@ -8,6 +8,7 @@
 * LayoutAmountInfo and LayoutSideInfo are renamed to AmountInfo and SideInfo
 * View.tag() is simplified
 * Overriding View.className is optional. It will return the class name correctly.
+* View.fellows returns a map instead of a collection, and View.getFellow is removed.
 
 **Features:**